+++ src/env/browser/platform.ts
@@ -29,6 +29,16 @@ export function getAltKeySymbol(): string {
 	return 'Alt';
 }
 
+export function getCmdKeySymbol(): string {
+	if (isMac) return '⌘';
+	return 'Ctrl';
+}
+
+export function getShiftKeySymbol(): string {
+	if (isMac) return '⇧';
+	return 'Shift';
+}
+
 /**
  * Returns an identifier for the current remote instance.
  * Not applicable in browser environment.